macro-invoc-lexer: Split implementation in its own file
